Move function to ProductList / HomeScreen
